init autotool test
