+++ Makefile.am
@@ -1,0 +1,3 @@
+bin_PROGRAMS		= can_test
+can_test_SOURCES	= can.c can.h can_test.c
+can_test_CPPFLAGS	= -I include/
