+++ stamp-h1
@@ -1,0 +1,1 @@
+timestamp for config.h
